Feature/multibranch scan (#372) (#374)
* Added functionality for Multibranch Scan

Functionality added to trigger a Jenkins Multibranch scan when a branch is added or deleted in bitbucket.

* Added functions for data capture

Added a function to capture both the from Hash and change type items that are inside of the payload received from Bitbucket

* Added full functions for payload data

Functions added get the from hash and change type items from the Bitbucket payload

* Added constant variables

Added constant variables for the change type values along with the value of an empty hash.  These are used to determine when a branch is being added or deleted.

* Update repository URL

Updating the plugin repository to have the correct URL for a successful Maven Build

* Update Logic for Multibranch Pipeline Scan Trigger

Previously when triggering the scan it was done so for all pipelines. New logic ensures that it only occurs for the repo captured from the payload if it matches the branch source specified in the multibranch pipeline

* Adding functions to pull clone links from payload

Functiions have been added to pull http and ssh clone links from bitbucket payload

* Added functions for capturing clone links

Added Functions to pull the SSH and HTTP clone links from the Bitbucket payload

* Update BitBucketPPRJobProbe.java

Ran into null pointer issue causing issues with other triggers.  Updates made to resolve null pointer occurrence.

Co-authored-by: s78258819-svg <s78258819@gmail.com>